handle reboot_required; make changes take effect immediately
Added one new input parameter and one output variable.
`kernel_settings_reboot_ok` - used to tell the role if it is ok
for the role to reboot the machine if the changes need a reboot
to be applied (e.g. bootloader cmdline settings).

The role will export `kernel_settings_reboot_required: true` if
the user did not set `kernel_settings_reboot_ok: true` and the
changes need a reboot to be applied.  The user is responsible
for rebooting the machine when convenient.

This also changes the handler behavior - now the handler only
does the reboot - the kernel_settings changes are now applied
immediately, by restarting tuned.  This makes the
kernel_settings role more like the `sysctl` module, which
applies the changes immediately.  If the kernel_settings
module is called from other roles e.g. a role that deploys
Elasticsearch, the changes have to be applied immediately,
rather than waiting until the end of the play containing
the roles, when the handler is executed, or Elasticsearch
may not be able to start.  I also didn't want to make it a
requirement to force the user to use `meta: flush_handlers`
after calling the role.

This change also uses the role parameters explicitly in the
test, to make it more obvious to test writers what parameters
there are and how they are specified.

On Fedora, due to the problem with the cloud images used for CI,
we need to disable bootloader cmdline testing, but we still want
to test everything else, so we disable the variables used to do
the bootloader cmdline testing on Fedora.