module should indicate if reboot needed to apply changes
Adds a new return value from the module `reboot_required` - this is
a boolean value, `false` by default.  If `true`, this means there
was a change made that will require rebooting the managed host in
order for the change to take effect.
I added a `RETURN` section to document the module return values.
I cleaned up the formatting of the examples and of the return
messages.