use panic instead of quiet in blcmd examples, tests
Some of the systems we use for testing have `quiet` as a built-in
bootloader cmdline value.  The kernel_settings role cannot remove
or replace built-in values.  Instead, use `panic` with a value
of `10001` as it is seldom used and applicable to all platforms.
This also gives us a non-string value to use for testing.
There was a bug with using non-string values in blcmdline
values.  This commit also fixes that bug.