replace bootloader cmdline bool values with on/off
The use of `on` and `off` for bootloader cmdline values is
quite common - https://www.kernel.org/doc/html/latest/admin-guide/kernel-parameters.html

Even though we will strongly encourage users to use a YAML string
value for bootloader cmdline settings, it will help users avoid
mistakes if we convert YAML boolean values to `"on"`/`"off"`.

For example, if the user types in this instead of `value: "on"`:
```yaml
kernel_settings_bootloader_cmdline:
  - name: someparam
    value: on
```
we will convert this to `someparam=on`.  This also means values
like:
```yaml
kernel_settings_bootloader_cmdline:
  - name: someparam
    value: true
```
will also be converted to `someparam=on`.