do not use omit - does not work on nested values with ansible 2.7
The use of `omit` with second level or lower values does not
work on ansible 2.7.  The keys are passed to the module with
the value `__omit_token_xxxx`.  Instead, use the value `none`
or some other value which is empty appropriate for the key being
set, and have the module strip all empty values.