use import_role instead of include_vars for role vars
The test verification needs access to the role private vars from
vars/main.yml.  Using include_vars requires a path which may be
different depending on the packaging.  Instead, rely on a "trick"
with import_role - using `when: false` will include the role, skipping
the explicit tasks, but running the implicit tasks such as loading
the vars and defaults.